requirements.txt
python-dotenv
web3==6.9.0
gql[requests]==3.4.1
pycoingecko==3.1.0

git+https://github.com/BalancerMaxis/bal_addresses@0.8.4
